shared memory timer
