+++ example/multikernel_timer/timer.system
@@ -7,14 +7,20 @@
 <system>
     <memory_region name="timer" size="0x10_000" phys_addr="0xffd0f000" />
 
+    <memory_region name="shared" size="0x1000" />
+
     <protection_domain name="timer">
         <program_image path="timer.elf" />
         <map mr="timer" vaddr="0x2_000_000" perms="rw" cached="false" setvar_vaddr="timer_regs" />
         <irq irq="42" id="0" trigger="edge" />
+
+        <map mr="shared" vaddr="0x2_100_000" perms="rw" cached="false" setvar_vaddr="symbol_shared_buffer" />
     </protection_domain>
 
     <protection_domain name="client" core="1">
         <program_image path="client.elf" />
+
+        <map mr="shared" vaddr="0x2_100_000" perms="rw" cached="false" setvar_vaddr="symbol_shared_buffer" />
     </protection_domain>
 
     <channel>
